requirements.txt
langchain-community 
langchainhub 
langgraph 
psycopg2 
python-dotenv  
langchain
jupyte